+++ tests/Pest.php
@@ -4,4 +4,4 @@ declare(strict_types=1);
 
 use AmazScript\Einvoicing\Tests\TestCase;
 
-uses(TestCase::class)->in('Feature', 'Unit');
+uses(TestCase::class)->in('Feature', 'Unit', 'Integration');